+++ vcontrold/client.h
@@ -12,10 +12,10 @@
 
 typedef struct txRx *trPtr;
 
-int recvSync(int fd,char *wait,char **recv); 
+ssize_t recvSync(int fd,char *wait,char **recv);
 int connectServer(char *host);
 void disconnectServer(int sockfd);
-int sendServer(int fd,char *s_buf, int len);
+size_t sendServer(int fd,char *s_buf, size_t len);
 trPtr sendCmdFile(int sockfd,char *tmpfile);
 trPtr sendCmds(int sockfd,char *commands);
